+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 # PEST IntelliJ Changelog
 
 ## [Unreleased]
+
+## [1.3.0]
 ### Fixed
 - Changed services to light services for auto disposable
 - Fixed null pointer error when no virtual file
@@ -12,7 +14,7 @@
 
 ### Added
 - Added higher order expectation type provider
-- Added support for xdebug3 and xdebug2 coverage option 
+- Added support for xdebug3 and xdebug2 coverage option
 
 ## [1.2.2]
 ### Fixed